chore: fixed divergent files in compas openscd
