--- packages/compas-open-scd/test/mock-editor-logger.ts
@@ -1,7 +1,0 @@
-import { LitElement, customElement } from 'lit-element';
-
-import { Editing } from 'open-scd/src/Editing.js';
-import { Historing } from '../src/Historing.js';
-
-@customElement('mock-editor-logger')
-export class MockEditorLogger extends Editing(Historing(LitElement)) {}
