+++ packages/compas-open-scd/test/unit/locamation/LocamationIEDList.test.ts
@@ -1,14 +1,17 @@
-import {expect, fixture, html} from "@open-wc/testing";
+import { expect, fixture, html } from '@open-wc/testing';
 
-import {MockWizard} from "open-scd/test/mock-wizard.js";
-import {fetchDoc} from "../wizards/test-support.js";
-import {initializeNsdoc, Nsdoc} from "open-scd/src/foundation/nsdoc.js";
+import { MockWizard } from 'open-scd/test/mock-wizard.js';
+import { fetchDoc } from 'open-scd/test/unit/wizards/test-support.js';
+import { initializeNsdoc, Nsdoc } from 'open-scd/src/foundation/nsdoc.js';
 
-import "open-scd/test/mock-wizard.js";
+import 'open-scd/test/mock-wizard.js';
 
-import {LocamationIEDListElement, locamationIEDListWizard} from "../../../src/locamation/LocamationIEDList.js";
+import {
+  LocamationIEDListElement,
+  locamationIEDListWizard,
+} from '../../../src/locamation/LocamationIEDList.js';
 
-import "../../../src/locamation/LocamationIEDList.js";
+import '../../../src/locamation/LocamationIEDList.js';
 
 describe('Wizards for Locamation Plugin to select the IED/Logical Device - ', () => {
   let nsdoc: Nsdoc;
@@ -41,7 +44,12 @@ describe('Wizards for Locamation Plugin to select the IED/Logical Device - ', ()
     beforeEach(async () => {
       doc = await fetchDoc('/test/testfiles/locamation/LCMTN_VMU_MMS.scd');
 
-      element = await fixture(html`<locamation-ied-list .doc="${doc}" .nsdoc="${nsdoc}"></locamation-ied-list>`);
+      element = await fixture(
+        html`<locamation-ied-list
+          .doc="${doc}"
+          .nsdoc="${nsdoc}"
+        ></locamation-ied-list>`
+      );
       await element.requestUpdate();
     });
 
@@ -56,7 +64,12 @@ describe('Wizards for Locamation Plugin to select the IED/Logical Device - ', ()
     beforeEach(async () => {
       doc = await fetchDoc('/test/testfiles/valid2007B4.scd');
 
-      element = await fixture(html`<locamation-ied-list .doc="${doc}" .nsdoc="${nsdoc}"></locamation-ied-list>`);
+      element = await fixture(
+        html`<locamation-ied-list
+          .doc="${doc}"
+          .nsdoc="${nsdoc}"
+        ></locamation-ied-list>`
+      );
       await element.requestUpdate();
     });
