+++ packages/compas-open-scd/test/unit/mock-compas-plugger.ts
@@ -3,5 +3,7 @@ import { Plugging } from '../../src/Plugging.js';
 import { Editing } from 'open-scd/src/Editing.js';
 import { Historing } from '../../src/Historing.js';
 
-@customElement('mock-plugger')
-export class MockPlugger extends Plugging(Editing(Historing(LitElement))) {}
+@customElement('mock-compas-plugger')
+export class MockCompasPlugger extends Plugging(
+  Editing(Historing(LitElement))
+) {}
