--- packages/compas-open-scd/test/unit/mock-logger.ts
@@ -1,5 +1,0 @@
-import { LitElement, customElement } from 'lit-element';
-import { Historing } from '../../src/Historing.js';
-
-@customElement('mock-logger')
-export class MockLogger extends Historing(LitElement) {}
