--- packages/compas-open-scd/test/unit/mock-setter.ts
@@ -1,5 +1,0 @@
-import { LitElement, customElement } from 'lit-element';
-import { Setting } from '../../src/Setting.js';
-
-@customElement('mock-setter')
-export class MockSetter extends Setting(LitElement) {}
